Example Project Demonstrating Problem
This project highlights an issue with using the Mox library. This
problem may occur with other mocking libraries, but we're using Mox so
that's why this example does too.

The problem can be briefly summarized as follows:

When a GenServer is added to the static application supervision tree, it
will be started by default. If this GenServer depends on a service which
you would like to replace with a Mock in tests, this is only possible if
your service is not invoked during the GenServer start up process, so
`start_link`, `init`, `handle_continue` etc. If you attempt to call your
service during any of these functions, whilst the code may work in
production, in tests when a mock is defined with `Mox.defmock` call, it
is already too late to define an expection.

There are "work arounds" of course, but none are particularly wholesome.

1. You could define your own TestMock module, which means it will be
defined at the time of application boot.
2. You could test for the `Mix.env` or `function_exported?` but uggg!!
3. You could call `mix test --no-start` and manually start.