Mix test --no-start
By preventing the application tree starting automatically we're able to
start the worker as we need, however this does mean each test will need
to start the processes it needs. This could be a contenious issue.

One downside of this approach is that we're now forced to set Mox into
"global" mode because we don't have a pid to pass to `Mox.allow/2` and
we need to call `Mox.expect/2` before calling `Worker.start_link/0` AND
we need the newly created `Worker` process to be able to "see" the
mocked `foo/0` function. Basically it's a chicken and egg situation.