Hand crafted test services
This version sets up services in the test by updating the environment
variable from within the test case.

To make the worker use the new service requires restarting the worker.

This approach isn't compatible with Mox, and the assertion is based on
values hard coded in the custom test services which feels kind of
clunky.